+++ lib/davinci_pas_test_kit/custom_groups/v2.0.1/pas_client_denial_group.rb
@@ -1,7 +1,7 @@
 require_relative 'client_tests/pas_client_denial_submit_test'
 require_relative 'client_tests/pas_client_response_attest'
-require_relative 'client_tests/pas_client_pas_response_bundle_validation_test'
-require_relative 'client_tests/pas_client_pas_request_bundle_validation_test'
+require_relative 'client_tests/pas_client_response_bundle_validation_test'
+require_relative 'client_tests/pas_client_request_bundle_validation_test'
 require_relative '../../user_input_response'
 require_relative '../../tags'
